# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: KIWAME API REST
  description: API for the management of KIWAME 
...
securityDefinitions:
  ApiKey:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
    description: "Security sheme migrated from security policy"
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.


     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.


     The following syntax must be used in the 'Authorization' header :

       Bearer xxxxxx.yyyyyyy.zzzzzz


# during dev, should point to your local machine
#host: 80.58.61.69  #PRODUCCIÓN
#host: 138.4.47.33:8163  #UPM DEV
#host: localhost:8090  #LOCAL DEV
# basePath prefixes all resource paths
basePath: / #dds
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /rule/{id}:
    x-swagger-router-controller: rule
    get:
      tags:
        - rules
      summary: Toogle enabled 
      description: toogle enable/disabled rule
      operationId: findById
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to toogle
          type: string
          required: true
      responses:
        "200":
          description: rule retrieves
          schema:
            $ref: "#/definitions/RuleResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Rule Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    put:
      tags:
        - rules
      summary: Toogle enabled 
      description: toogle enable/disabled rule
      operationId: toogleEnabled
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
        - application/octet-stream
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to toogle
          type: string
          required: true
      responses:
        "200":
          description: rule Enabled/Disabled
          schema:
            $ref: "#/definitions/ServerResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Rule Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    delete:
      tags:
        - rules
      summary: Delete rule
      description: Delete rule
      operationId: deleteRule
      security:
        - Bearer: []
      x-security-scopes:
        - profesional
        - admin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the rule to delete
          type: string
          required: true
      responses:
        "200":
          description: Rule removed
          schema:
            $ref: "#/definitions/ServerResponse"
        "403":
          description: Access Denied
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: Profile Not Found
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
  /drink:
    x-swagger-router-controller: drink
    post:
      tags:
        - drinks
      operationId: newDrink
      summary: post a  drink
      description: Upload a drink for the logged user
      security:
        - Bearer: []
      x-security-scopes:
        - patient
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - name:  drink
          in: body
          description: The drink JSON you want to POST
          schema:
            $ref: "#/definitions/DrinkObject"
          required: true

      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/ServerResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /drink/history/{user}:
    x-swagger-router-controller: drink
    get:
      tags:
        - drinks
      operationId: getDrinks
      summary: Retrieve drink history
      description: Retrieve the drink history of the user
      security:
        - Bearer: []
      x-security-scopes:
        - patient
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - name: user
          in: path
          description: The nickname of the user to retrieve drink history
          type: string
          required: true
        - name: next
          in: query
          description: the id for retrieve the next page of results
          type: number
          required: false
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/DrinkHistoryResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/RepeatedResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /drink/types:
    x-swagger-router-controller: drink
    get:
      tags:
        - drinks
      operationId: getDrinkTypes
      summary: Retrieve drink types
      description: Retrieve the drink types
      security:
        - Bearer: []
        - ApiKey: []
      x-security-scopes:
        - admin
        - patient
        - profesional
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - description: Provide a header to specify the language.
          in: header
          name: Accept-Language
          required: false
          type: string  
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/DrinkTypesResponse"
        "403":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/AccessDeniedResponse"
        "404":
          description: "User not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
definitions:
...
  DrinkObject:
    required:
      - type
      - volume
    properties:
      type:
        type: number
        example: 1
      volume:
        type: number
        example: 1
      created:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
  DrinkHistoricObject:
    required:
      - type
      - volume
      - created
    properties:
      type:
        $ref: "#/definitions/DrinkTypeObject"
      volume:
        type: number
        example: 1
      created:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
  DrinkTypeObject:
    required:
      - type
      - name
      - nutrients
    properties:
      type:
        type: number
        example: 1
      name:
        type: string
        example: water
      nutrients:
         $ref: "#/definitions/NutrientObject"
...
  RuleRetrievedObject:
    required:
      - type
      - condition
      - value
      - notificationType
    properties:
      id:
        type: string
      patient:
        type: string
        example: nickname
      type:
        type: number
        example: 1
      measureType:
        type: number
        example: 2
      nutrientType:
        type: number
        example: 1
      condition:
        type: number
        example: 1
      value:
        type: number
        example: 10
      period:
        type: number
        minimum: 1
        maximum: 30
        example: 1
      notificationType:
        type: number
        example: 1
      enabled:
        type: boolean
        example: true
      created:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
      updated:
        type: string
        format: date-time
        example: "2016-06-14T10:24:11.387Z"
...
  NutrientObject:
      required:
        - total_fat
        - saturated_fat
        - cholesterol
        - sodium
        - total_carbohydrate
        - dietary_fiber
        - sugars
        - protein
        - potassium
        - phosphorus
        - water
        - alcohol
        - caffeine
        - calcium
        - iron
      properties:
        total_fat:
          type: number
          example: 1
        saturated_fat:
          type: number
          example: 1
        cholesterol:
          type: number
          example: 1
        sodium:
          type: number
          example: 1
        total_carbohydrate:
          type: number
          example: 1
        dietary_fiber:
          type: number
          example: 1
        sugars:
          type: number
          example: 1
        protein:
          type: number
          example: 1
        potassium:
          type: number
          example: 1
        phosphorus:
          type: number
          example: 1
        water:
          type: number
          example: 1
        alcohol:
          type: number
          example: 1
        caffeine:
          type: number
          example: 1
        calcium:
          type: number
          example: 1
        iron:
          type: number
          example: 1
        calories:
          type: number
          example: 1
...
  NotFoundResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Error: Resource Not found"
      statusCode:
        type: number
        example: 404
...
  DrinkHistoryResponse:
    required:
      - message
      - statusCode
      - drinks
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      next:
        type: string
        example: "drinks/history/patient?next=2"
      drinks:
        type: array
        items:
          $ref: "#/definitions/DrinkHistoricObject"
  DrinkTypesResponse:
    required:
      - message
      - statusCode
      - drinks
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      drinks:
        type: array
        items:
          $ref: "#/definitions/DrinkTypeObject"
...
  RuleResponse:
    required:
      - message
      - statusCode
      - rule
    properties:
      message:
        type: string
        example: "success"
      statusCode:
        type: number
        example: 200
      rule:
          $ref: "#/definitions/RuleRetrievedObject"
  ServerResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Success"
      statusCode:
        type: number
        example: 200
...
  BadRequestResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Bad Request"
      statusCode:
        type: number
        example: 400
  AccessDeniedResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Access Denied"
      statusCode:
        type: number
        example: 403
  RepeatedResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "item already exists"
      statusCode:
        type: number
        example: 409
...
  ServerErrorResponse:
    required:
      - message
      - statusCode
    properties:
      message:
        type: string
        example: "Error: Internal Server Error"
      statusCode:
        type: number
        example: 500